move new site to individual extension
