--- multisrc/overrides/madara/wickedwitchscan/additional.gradle
@@ -1,5 +1,0 @@
-dependencies {
-    implementation(project(":lib:cryptoaes"))
-    implementation(project(":lib:randomua"))
-    compileOnly("com.github.tachiyomiorg:image-decoder:fbd6601290")
-}
