--- multisrc/overrides/madara/wickedwitchscan/src/WickedWitchScanFactory.kt
@@ -1,10 +1,0 @@
-package eu.kanade.tachiyomi.extension.pt.wickedwitchscan
-
-import eu.kanade.tachiyomi.source.SourceFactory
-
-class WickedWitchScanFactory : SourceFactory {
-    override fun createSources() = listOf(
-        WickedWitchScan(),
-        WickedWitchScanNew(),
-    )
-}
