+++ src/pt/wickedwitchscannovo/res/mipmap-hdpi/ic_launcher.png
