+++ src/pt/wickedwitchscannovo/res/mipmap-mdpi/ic_launcher.png
