+++ src/pt/wickedwitchscannovo/res/mipmap-xhdpi/ic_launcher.png
